fix: uint256 counter

+++ contracts/facets/InstantActions/InstantActionsFacetImpl.sol
@@ -200,7 +200,7 @@ library InstantActionsFacetImpl {
 		intentId = ++intentLayout.lastOpenIntentId;
 		uint256 tradeId = ++intentLayout.lastTradeId;
 
-		uint32 counter = 0; // Initialize counter
+		uint256 counter = 0; // Initialize counter
 		bytes memory userDataWithCounter = LibIntent.addCounter(signedOpenIntent.userData, counter);
 
 		address[] memory partyBsWhitelist = new address[](1);
